Move startup telemetry code to its own module.
